requirements.txt
Flask
python-dotenv
google-generativeai
requests
Flask-CORS
# Add Flask-SSE and redis if you are using the SSE version
# Flask-SSE
# redis
google-cloud-bigquery